+++ NEWS.md
@@ -1,5 +1,7 @@
 # broom.helpers (development version)
 
+* New function `tidy_select_variables()` to keep/drop
+  selected variables in the output (#45)
 * Bug fix for non standard variable names containing
   a character that would have a special meaning in
   a regular expression (#44)